Color Scheme: Light Mode and Dark Mode
This is a WIP.

Adding the ability to switch to a light mode. I don't know why anyone would want it but why not.

Created a new "light" class that can be appended to the 'body' element or any 'a' elements. The class is appended by clicking a "Toggle Color Scheme" link which I'll later change to be a fancy looking icon or something. It has some